+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 6.3
 - Uncheck Saturday & Sunday for Recreate Calendar too in Calendar.php
 - HTML Minutes: use "number" input type in Calendar.php
 - SQL add Donate link in rosariosis.sql, rosariosis_es.sql & rosariosis_fr.sql
+- HTML smaller size for 'Displaying %d through %d' in ListOutput.fnc.php
 
 Changes in 6.2.3
 ----------------